# src/main/java/podobnyi/dev/event_manager/web/ErrorMessageResponse.java
package podobnyi.dev.event_manager.web;

import java.time.LocalDateTime;

public record ErrorMessageResponse(
        String message,
        String detailedMessage,
        LocalDateTime dateTime
) {
}